# backend/openapi/openapi.json
{
  "components": {
    "schemas": {
      "AccessRuleCreate": {
        "additionalProperties": false,
        "properties": {
          "effect": {
            "enum": [
              "ALLOW",
              "DENY"
            ],
            "title": "Effect",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "inherits": {
            "default": false,
            "title": "Inherits",
            "type": "boolean"
          },
          "permission": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Permission",
            "type": "string"
          },
          "principal_id": {
            "exclusiveMinimum": 0.0,
            "title": "Principal Id",
            "type": "integer"
          },
          "principal_type": {
            "enum": [
              "USER",
              "GROUP"
            ],
            "title": "Principal Type",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "principal_type",
          "principal_id",
          "permission",
          "effect"
        ],
        "title": "AccessRuleCreate",
        "type": "object"
      },
      "AccessRuleOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "title": "Created By",
            "type": "integer"
          },
          "effect": {
            "title": "Effect",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "inherits": {
            "title": "Inherits",
            "type": "boolean"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "permission": {
            "title": "Permission",
            "type": "string"
          },
          "principal_id": {
            "title": "Principal Id",
            "type": "integer"
          },
          "principal_type": {
            "title": "Principal Type",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "scope_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Id"
          },
          "scope_type": {
            "title": "Scope Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "principal_type",
          "principal_id",
          "permission",
          "scope_type",
          "scope_id",
          "effect",
          "inherits",
          "is_active",
          "expires_at",
          "reason",
          "created_by",
          "created_at"
        ],
        "title": "AccessRuleOut",
        "type": "object"
      },
...
      "EffectiveAccessOut": {
        "properties": {
          "allowed": {
            "title": "Allowed",
            "type": "boolean"
          },
          "ancestry": {
            "items": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "integer"
                  },
                  {
                    "type": "string"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "object"
            },
            "title": "Ancestry",
            "type": "array"
          },
          "matched_rule": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EffectiveAccessRule"
              },
              {
                "type": "null"
              }
            ]
          },
          "permission": {
            "title": "Permission",
            "type": "string"
          },
          "policy_version": {
            "title": "Policy Version",
            "type": "integer"
          },
          "reason_code": {
            "title": "Reason Code",
            "type": "string"
          },
          "scope_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Id"
          },
          "scope_type": {
            "title": "Scope Type",
            "type": "string"
          },
          "user_id": {
            "title": "User Id",
            "type": "integer"
          }
        },
        "required": [
          "user_id",
          "permission",
          "scope_type",
          "scope_id",
          "allowed",
          "reason_code",
          "policy_version",
          "ancestry"
        ],
        "title": "EffectiveAccessOut",
        "type": "object"
      },
      "EffectiveAccessRule": {
        "properties": {
          "effect": {
            "title": "Effect",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "inherits": {
            "title": "Inherits",
            "type": "boolean"
          },
          "principal_id": {
            "title": "Principal Id",
            "type": "integer"
          },
          "principal_type": {
            "title": "Principal Type",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "scope_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope Id"
          },
          "scope_type": {
            "title": "Scope Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "principal_type",
          "principal_id",
          "effect",
          "inherits",
          "reason",
          "scope_type",
          "scope_id"
        ],
        "title": "EffectiveAccessRule",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "AI-based Digital Document Management System — local vertical slice.",
    "title": "XENIUS DocVault API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/access/effective": {
      "get": {
        "operationId": "explain_effective_access_api_v1_access_effective_get",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "User Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "permission",
            "required": false,
            "schema": {
              "default": "VIEW",
              "title": "Permission",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "default": "DOC",
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scope Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EffectiveAccessOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Explain Effective Access",
        "tags": [
          "access"
        ]
      }
    },
    "/api/v1/access/rules/{rule_id}": {
      "delete": {
        "operationId": "revoke_access_rule_api_v1_access_rules__rule_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Rule Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessRuleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Revoke Access Rule",
        "tags": [
          "access"
        ]
      },
      "patch": {
        "operationId": "update_access_rule_api_v1_access_rules__rule_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Rule Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessRuleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessRuleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Access Rule",
        "tags": [
          "access"
        ]
      }
    },
    "/api/v1/access/{scope}/rules": {
      "post": {
        "operationId": "grant_global_access_rule_api_v1_access__scope__rules_post",
        "parameters": [
          {
            "in": "path",
            "name": "scope",
            "required": true,
            "schema": {
              "title": "Scope",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessRuleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessRuleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Grant Global Access Rule",
        "tags": [
          "access"
        ]
      }
    },
    "/api/v1/access/{scope}/{scope_id}/rules": {
      "get": {
        "operationId": "list_access_rules_api_v1_access__scope___scope_id__rules_get",
        "parameters": [
          {
            "in": "path",
            "name": "scope",
            "required": true,
            "schema": {
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "scope_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Scope Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AccessRuleOut"
                  },
                  "title": "Response List Access Rules Api V1 Access  Scope   Scope Id  Rules Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Access Rules",
        "tags": [
          "access"
        ]
      },
      "post": {
        "operationId": "grant_access_rule_api_v1_access__scope___scope_id__rules_post",
        "parameters": [
          {
            "in": "path",
            "name": "scope",
            "required": true,
            "schema": {
              "title": "Scope",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "scope_id",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Scope Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessRuleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessRuleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Grant Access Rule",
        "tags": [
          "access"
        ]
      }
    },
...
  }
}